Move cfssl client to the wider scope
